+++ lib/src/ui/views/steps/customize/customize_view.dart
@@ -1,5 +1,3 @@
-import 'dart:io';
-
 import 'package:flutter/material.dart';
 import 'package:flutter_installer/src/app/models/user_choice.model.dart';
 import 'package:flutter_installer/src/ui/global/app_colors.dart';
@@ -319,55 +317,6 @@ class CustomizeView extends StatelessWidget {
                     Row(
                       mainAxisAlignment: MainAxisAlignment.spaceAround,
                       children: <Widget>[
-                        Platform.isMacOS
-                            ? Container(
-                                width: blockSize(context) * 30,
-                                child: TextField(
-                                  controller: model.sudoPasswordController,
-                                  decoration: InputDecoration(
-                                    border: OutlineInputBorder(
-                                      borderRadius: BorderRadius.circular(
-                                        20,
-                                      ),
-                                    ),
-                                    labelText: 'Sudo Password',
-                                    labelStyle: TextStyle(
-                                      fontFamily: 'RobotoMono',
-                                      fontSize: blockSize(context) * 1.5,
-                                    ),
-                                    hintText: 'Please Enter Your Sudo Password',
-                                    hintStyle: TextStyle(
-                                      fontFamily: 'RobotoMono',
-                                      fontSize: blockSize(context) * 1.1,
-                                      color: model.sudoPasswordTextFieldHasError
-                                          ? dangerColor
-                                          : lynchColor,
-                                    ),
-                                    suffixIcon: IconButton(
-                                      icon: Icon(
-                                        model.obscureSudoPassword
-                                            ? Icons.visibility
-                                            : Icons.visibility_off,
-                                      ),
-                                      onPressed: () {
-                                        model.setobscureSudoPassword(
-                                            !model.obscureSudoPassword);
-                                      },
-                                    ),
-                                  ),
-                                  style: TextStyle(
-                                    fontFamily: 'RobotoMono',
-                                    fontSize: blockSize(context) * 1.1,
-                                    // fontWeight: FontWeight.bold,
-                                    color: model.sudoPasswordTextFieldHasError
-                                        ? dangerColor
-                                        : lynchColor,
-                                  ),
-                                  onChanged: model.setSudoPassword,
-                                  obscureText: model.obscureSudoPassword,
-                                ),
-                              )
-                            : Container(),
                         CustomButton(
                           text: 'Next',
                           textStyle: TextStyle(
@@ -396,24 +345,6 @@ class CustomizeView extends StatelessWidget {
                             }
                             model.setChooseFolderTextFieldHasError(false);
 
-                            if (Platform.isMacOS) {
-                              if (model.sudoPasswordController.text == null ||
-                                  model.sudoPassword == null ||
-                                  model.sudoPasswordController.text.trim() ==
-                                      '' ||
-                                  model.installationPath.trim() == '') {
-                                model.setSudoPasswordTextFieldHasError(true);
-
-                                model.showSnackBar(
-                                  title: 'Error Occured',
-                                  message: 'You must put your Sudo password!',
-                                );
-
-                                return;
-                              }
-                            }
-                            model.setSudoPasswordTextFieldHasError(false);
-
                             onNextPressed(
                               UserChoice(
                                 installationPath: model.installationPath,
@@ -424,8 +355,6 @@ class CustomizeView extends StatelessWidget {
                                 installIntelliJIDEA: model.installIntelliJIDEA,
                                 installGit: model.installGit,
                                 flutterChannel: model.flutterChannel,
-                                sudoPassword:
-                                    Platform.isMacOS ? model.sudoPassword : '',
                               ),
                             );
                           },