+++ lib/src/ui/views/steps/customize/customize_view_model.dart
@@ -13,7 +13,6 @@ class CustomizeViewModel extends CustomBaseViewModel {
   final LocalStorageService _localStorageService =
       locator<LocalStorageService>();
   final TextEditingController chooseFolderController = TextEditingController();
-  final TextEditingController sudoPasswordController = TextEditingController();
 
   bool _showAdvanced = false;
   bool get showAdvanced => _showAdvanced;
@@ -22,24 +21,10 @@ class CustomizeViewModel extends CustomBaseViewModel {
     notifyListeners();
   }
 
-  bool _obscureSudoPassword = true;
-  bool get obscureSudoPassword => _obscureSudoPassword;
-  void setobscureSudoPassword(bool newValue) {
-    _obscureSudoPassword = newValue;
-    notifyListeners();
-  }
-
   bool _chooseFolderTextFieldHasError = false;
   bool get chooseFolderTextFieldHasError => _chooseFolderTextFieldHasError;
   void setChooseFolderTextFieldHasError(bool newValue) {
     _chooseFolderTextFieldHasError = newValue;
-    notifyListeners();
-  }
-
-  bool _sudoPasswordTextFieldHasError = false;
-  bool get sudoPasswordTextFieldHasError => _sudoPasswordTextFieldHasError;
-  void setSudoPasswordTextFieldHasError(bool newValue) {
-    _sudoPasswordTextFieldHasError = newValue;
     notifyListeners();
   }
 
@@ -82,13 +67,6 @@ class CustomizeViewModel extends CustomBaseViewModel {
   FlutterChannel get flutterChannel => _flutterChannel;
   void setFlutterChannel(FlutterChannel newValue) {
     _flutterChannel = newValue;
-    notifyListeners();
-  }
-
-  String _sudoPassword;
-  String get sudoPassword => _sudoPassword;
-  void setSudoPassword(String newValue) {
-    _sudoPassword = newValue;
     notifyListeners();
   }